Tests: Move fixture portal_mock into conftests.py
For reuse by setup tests.
